+++ .gitignore
@@ -37,3 +37,8 @@ yarn-error.log*
 
 # yarn
 .yarn
+
+
+*.tgz
+packages/**/dist/*
+node_modules/
