+++ .npmignore
@@ -1,0 +1,18 @@
+node_modules/
+apps/
+docs/
+src/
+.gitignore
+.turbo
+src/**/*.test.ts
+src/**/*.test.js
+.*.swp
+._*
+.DS_Store
+.git
+.hg
+.npmrc
+.lock-wscript
+.svn
+.wafpickle-*
+npm-debug.log
