+++ apps/docs/package.json
@@ -4,7 +4,6 @@
 	"private": true,
 	"scripts": {
 		"dev": "next dev",
-		"build": "next build",
 		"start": "next start",
 		"lint": "next lint"
 	},
