+++ package.json
@@ -1,40 +1,61 @@
 {
-	"name": "next-mobx-store",
-	"private": true,
-	"packageManager": "yarn@3.5.1",
-	"scripts": {
-		"build": "turbo run build",
-		"dev": "turbo run dev",
-		"lint": "turbo run lint",
-		"test": "turbo run test"
-	},
-	"devDependencies": {
-		"@babel/preset-env": "^7.21.5",
-		"@babel/preset-react": "^7.18.6",
-		"@babel/preset-typescript": "^7.21.5",
-		"@testing-library/react": "^14.0.0",
-		"@testing-library/user-event": "^14.4.3",
-		"@types/jest": "^29.5.1",
-		"@types/react": "^18.2.6",
-		"@types/react-dom": "^18.2.4",
-		"eslint": "^7",
-		"jest": "^29.5.0",
-		"ts-jest": "^29.1.0",
-		"turbo": "^1.9.3",
-		"typescript": "^5.0.4"
-	},
-	"workspaces": [
-		"apps/*",
-		"packages/*"
-	],
-	"dependencies": {
-		"immutable": "^4.3.0",
-		"json-immutable": "^0.5.0",
-		"mobx": "^6.9.0",
-		"mobx-react-lite": "^3.4.3",
-		"next": "^13.4.1",
-		"react": "^18.2.0",
-		"react-dom": "^18.2.0",
-		"tsup": "^6.7.0"
-	}
+  "name": "@casentino/next-mobx-store",
+  "private": true,
+  "packageManager": "yarn@3.5.1",
+  "scripts": {
+    "build": "turbo run build",
+    "dev": "turbo run dev",
+    "lint": "turbo run lint",
+    "test": "turbo run test"
+  },
+  "devDependencies": {
+    "@babel/preset-env": "^7.21.5",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/preset-typescript": "^7.21.5",
+    "@swc/core": "^1.3.60",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.5.1",
+    "@types/react": "^18.2.6",
+    "@types/react-dom": "^18.2.4",
+    "eslint": "^7",
+    "jest": "^29.5.0",
+    "ts-jest": "^29.1.0",
+    "turbo": "^1.9.3",
+    "typescript": "^5.0.4"
+  },
+  "workspaces": [
+    "apps/*",
+    "packages/*"
+  ],
+  "dependencies": {
+    "immutable": "^4.3.0",
+    "json-immutable": "^0.5.0",
+    "mobx": "^6.9.0",
+    "mobx-react-lite": "^3.4.3",
+    "next": "^13.4.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "tsup": "^6.7.0"
+  },
+  "description": "",
+  "version": "0.0.1",
+  "main": "babel.config.js",
+  "directories": {
+    "doc": "docs"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/casentino/next-mobx-store.git"
+  },
+  "keywords": [
+    "next",
+    "mobx"
+  ],
+  "author": "casentino",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/casentino/next-mobx-store/issues"
+  },
+  "homepage": "https://github.com/casentino/next-mobx-store#readme"
 }
