--- packages/next-mobx-store-core/index.ts
@@ -1,4 +1,0 @@
-export * from './src/createRootStore';
-export * from './src/hydrationUtils';
-export * from './src/RootStoreProvider';
-export { default as initializeRootStore } from './src/initializeRootStore';
