+++ packages/next-mobx-store-core/src/index.ts
@@ -1,0 +1,7 @@
+import createRootStore from './createRootStore';
+
+import { serializeStore, deserializeStore } from './hydrationUtils';
+
+import RootStoreProvider from './RootStoreProvider';
+
+export { RootStoreProvider, createRootStore, serializeStore, deserializeStore };
