+++ packages/next-mobx-store-core/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.json",  
+  "include": [
+    "src/**/*.ts"
+  ],  
+  "exclude": ["dist", "node_modules", "./tsup.config.ts"],
+}
