--- packages/next-mobx-store-hooks/index.ts
@@ -1,3 +1,0 @@
-export * from './src/useRootStore';
-export * from './src/useHydrateProps';
-export * from './src/useLocalHydrateStore';
