+++ packages/next-mobx-store-hooks/src/useLocalHydrateStore.ts
@@ -1,4 +1,4 @@
-import { deserializeStore } from '@next-mobx-store/core';
+import { deserializeStore } from '@next-mobx-store/core/src/hydrationUtils';
 import { LocalHydrateStores, LocalPageProps } from '@next-mobx-store/type';
 
 export default function useLocalHydrateStores<P extends LocalPageProps>(stores: LocalHydrateStores, props: P) {
