+++ packages/next-mobx-store-hooks/tsconfig.json
@@ -1,0 +1,5 @@
+{
+	"extends": "../../tsconfig.json",
+	"include": ["src/**/*.ts"],
+	"exclude": ["dist", "node_modules"]
+}
