+++ packages/next-mobx-store-type/package.json
@@ -1,7 +1,10 @@
 {
-  "name": "@next-mobx-store/type",
-  "version": "0.0.0",
-  "types": "./index.ts",
-  "main": "./index.ts",
-  "license": "MIT"
+	"name": "@next-mobx-store/type",
+	"version": "0.0.0",
+	"main": "./src/index.ts",
+	"module": "./src/index.ts",
+	"license": "MIT",
+	"scripts": {
+		"build": "tsc"
+	}
 }
