+++ packages/next-mobx-store-type/src/HydrationType.ts
@@ -1,6 +1,6 @@
 import { ObservableMap, ObservableSet } from 'mobx';
 import { IHydrationStore } from './IHydrationStore';
-import { ExcludeType, TransformMap, TransformSet, ToNotOptional, IncludeType } from './Utils';
+import { ExcludeType, TransformMap, TransformSet, ToNotOptional } from './Utils';
 
 export type HydrationDataType<Store extends object = object> = TranshFormHydrationStore<
 	Pick<
