+++ packages/next-mobx-store-type/src/HydrationUtilType.ts
