+++ packages/next-mobx-store-type/src/IHydrationStore.ts
