+++ packages/next-mobx-store-type/src/IRootStore.ts
