+++ packages/next-mobx-store-type/src/Utils.ts
