+++ packages/next-mobx-store-type/tsconfig.json
@@ -1,0 +1,9 @@
+{  
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "declaration": true,
+    "emitDeclarationOnly": true ,
+    "outDir": "./dist",
+  },
+  "include": ["src/**/*.ts"],  
+}
