+++ tsconfig.json
@@ -1,26 +1,21 @@
 {
-  "compilerOptions": {    
-    "target": "es5",
-    "lib": [
-      "DOM",
-      "DOM.Iterable",
-      "ESNext"
-    ] ,
-    "jsx": "preserve",
-    "module": "commonjs",
-    
-    "moduleResolution": "node",
-    "resolveJsonModule": true ,
-    "allowJs": true ,
-    "declaration": true,
-    "declarationMap": true,
-    "sourceMap": true, 
-    "noEmit": true ,
-    "isolatedModules": true,
-    "esModuleInterop": true ,
-    "forceConsistentCasingInFileNames": true ,
-    "strict": true ,
-    "skipLibCheck": true
-  },
+  "compilerOptions": {
+		"target": "esnext",
+		"lib": ["DOM", "DOM.Iterable", "ESNext"],
+		"jsx": "preserve",
+		"module": "commonjs",
+		"checkJs": true,
+		"moduleResolution": "node",
+		"resolveJsonModule": true,
+		"allowJs": true,
+		"declaration": true,
+		"declarationMap": false,
+		"sourceMap": false,
+		"isolatedModules": true,
+		"esModuleInterop": true,
+		"forceConsistentCasingInFileNames": true,
+		"strict": true,
+		"skipLibCheck": true,
+	},
   "exclude": ["node_modules"]
 }
